extract DefaultHttpProxyServerBootstrap as an external class, to fix the TODO comment.
